Thesaurus: Add copyright, file reformatted with black
